Kept the main function from deleting files when running PreRunInit()

+++ main.go
@@ -11,13 +11,27 @@ const (
 	defaultAuth string = "admin|password"
 )
 
+// Checking if a file exists
+func exists(path string) bool {
+	if _, err := os.Stat(path); os.IsNotExist(err) {
+		return false
+	}
+
+	return true
+}
+
 // Functions initialize the project structure
 func PreRunInit() {
 	os.Mkdir(config.DataDirectory, 775)
 	os.Mkdir(config.PostsDir, 775)
 
-	ioutil.WriteFile(config.AuthLoc, []byte(defaultAuth), 664)
-	ioutil.WriteFile(config.PostIndexLoc, []byte("0"), 664)
+	if !exists(config.AuthLoc) {
+		ioutil.WriteFile(config.AuthLoc, []byte(defaultAuth), 664)
+	}
+
+	if !exists(config.PostIndexLoc) {
+		ioutil.WriteFile(config.PostIndexLoc, []byte("0"), 664)
+	}
 }
 
 // The main function
